Move AFD retry logic up to ODSP driver.
